Compiler warnings - float conversion - double promotion

+++ examples/OpenGLWindow/SimpleCamera.cpp
@@ -298,16 +298,16 @@ void SimpleCamera::getCameraViewMatrix(float viewMatrix[16]) const
 
 void SimpleCamera::getCameraTargetPosition(double pos[3]) const
 {
-	pos[0] = m_data->m_cameraTargetPosition[0];
-	pos[1] = m_data->m_cameraTargetPosition[1];
-	pos[2] = m_data->m_cameraTargetPosition[2];
+	pos[0] = (double)m_data->m_cameraTargetPosition[0];
+	pos[1] = (double)m_data->m_cameraTargetPosition[1];
+	pos[2] = (double)m_data->m_cameraTargetPosition[2];
 }
 
 void SimpleCamera::getCameraPosition(double pos[3]) const
 {
-	pos[0] = m_data->m_cameraPosition[0];
-	pos[1] = m_data->m_cameraPosition[1];
-	pos[2] = m_data->m_cameraPosition[2];
+	pos[0] = (double)m_data->m_cameraPosition[0];
+	pos[1] = (double)m_data->m_cameraPosition[1];
+	pos[2] = (double)m_data->m_cameraPosition[2];
 }
 
 void SimpleCamera::getCameraTargetPosition(float pos[3]) const
